Minor reorganization and test fixes
